refactor : 폴더 정리 / Today's hot 수정
