+++ front/src/components/organisms/Home/BannerBox.vue
@@ -1,4 +1,5 @@
 <template>
+  <!--
   <v-row
     class="text-center"
     justify="center"
@@ -16,6 +17,24 @@
       </div>
     </v-col>
   </v-row>
+-->
+  <v-row
+    class="text-center"
+    justify="center"
+  >
+    <v-col
+      cols="12"
+      md="3"
+    >
+      <div class="test2">
+        <div
+          class="pa-4 secondary text-no-wrap rounded-pill grow-1"
+        >
+          <span class="hot">Today's Hot</span>
+        </div>
+      </div>
+    </v-col>
+  </v-row>
 </template>
 
 <script>
@@ -25,6 +44,13 @@ export default {
 </script>
 
 <style scoped>
+.test2{
+  width: auto;
+  height: 80px;
+  background-image: url("@/assets/배경1.jpg");
+  background-position: center;
+  background-size: 100%;
+}
 
 .today{
   width: auto;
