+++ front/src/components/organisms/ProductInfo/additional/ProductDetail.vue
