+++ front/src/components/organisms/ProductInfo/bar/ProductAppBar.vue
