+++ front/src/components/organisms/ProductInfo/footer/PageFooter.vue
